refactor:  페이지별 컴포넌트에서 공통 컴포넌트로 이동할 컴포넌트 정리 #102
